Move verify and parse tokens to AbstractOAuthClient
Closes #37660
